backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
python-multipart==0.0.6
openai==1.3.7
groq==0.4.1
google-cloud-texttospeech==2.14.1
python-dotenv==1.0.0
httpx==0.25.2
aiofiles==23.2.1